Revert "feat - landing page 제작중"
This reverts commit 165a68ba7168c87a95a4c029562b4285c69e9dd5.
